encuesta/requirements.txt
appdirs==1.4.3
Django==3.0.7
djangorestframework==3.11.0
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
django-cors-headers==3.7.0
Pillow==8.2.0
drf-extra-fields==3.1.1